% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{as_hinton_df}
\alias{as_hinton_df}
\alias{as_hinton_df.matrix}
\alias{as_hinton_df.data.frame}
\alias{as_hinton_df.table}
\alias{as_hinton_df.default}
\title{Convert an object to a tidy data frame for \code{geom_hinton()}}
\usage{
as_hinton_df(x, ...)

\method{as_hinton_df}{matrix}(x, ...)

\method{as_hinton_df}{data.frame}(
  x,
  rowname_col = "row",
  colname_col = "col",
  value_col = "weight",
  ...
)

\method{as_hinton_df}{table}(x, ...)

\method{as_hinton_df}{default}(x, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Additional arguments passed to the method.}

\item{rowname_col}{Name of the column that holds the row index. Default
\code{"row"}.}

\item{colname_col}{Name of the column that holds the column index. Default
\code{"col"}.}

\item{value_col}{Name of the column that holds the matrix values. Default
\code{"weight"}.}
}
\value{
A data frame suitable for use with \code{\link[=geom_hinton]{geom_hinton()}}.
}
\description{
Generic function that dispatches to a method appropriate for \code{x}.
Built-in methods exist for \code{matrix}, \code{data.frame}, and \code{table}.
}
\examples{
m <- matrix(c(1, -2, 3, -4), 2, 2)
as_hinton_df(m)

t <- table(c("a","b","a"), c("x","y","x"))
as_hinton_df(t)

}
