% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghinton-package.R
\docType{package}
\name{gghinton-package}
\alias{gghinton}
\alias{gghinton-package}
\title{gghinton: Hinton Diagrams for 'ggplot2'}
\description{
Provides a 'ggplot2' extension for drawing Hinton diagrams, a visualisation technique for numerical matrices in which the area of each square is proportional to the magnitude of the corresponding entry. For signed data, white squares indicate positive values and black squares indicate negative values on a grey background. Hinton diagrams are especially useful for visualising PCA weight matrices, correlation matrices, and transition matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robin-foster-rf/gghinton}
  \item Report bugs at \url{https://github.com/robin-foster-rf/gghinton/issues}
}

}
\author{
\strong{Maintainer}: Robin Foster \email{gghinton@proton.me} [copyright holder]

}
\keyword{internal}
